feat(ext/webgpu): add GPUCanvasContext to lazy loaded esm
